mp-hackncode-18-web/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography>=2.3
dj-database-url==0.5.0
Django==2.0.3
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
paypalrestsdk==1.13.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycparser==2.18
pyOpenSSL==17.5.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
